requirements.txt
Django==1.7
Markdown==2.4.1
MarkupSafe==0.18
South==1.0
django-rest-swagger==0.1.14
django-phonenumber-field==0.6
djangorestframework==2.4.2
django-cors-headers==0.13
phonenumbers==6.2.0
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
